Making some Updates
Making it run once within half a second when a page is first loaded.

Then afterwards, there is a 1 minute grace period you can still access the github issue page. After accessing it, the injection will work properly otherwise it won't make the change to the html. I think it might be due to the fact I'm using JQuery and jquery might need the page to actually be accessed before running properly.